Файл account/views.py разбит на несколько и они перенесены в папку views/
