+++ account/views/access.py
@@ -1,0 +1,124 @@
+"""
+----------------------------------------------
+
+Copyright © Egor Vavilov (Shecspi)
+Licensed under the Apache License, Version 2.0
+
+----------------------------------------------
+"""
+
+import logging
+
+from django.urls import reverse_lazy
+from django.contrib.auth import login
+from django.contrib.auth.models import User
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.shortcuts import render, redirect
+from django.views.generic import CreateView
+from django.contrib.auth.views import LoginView, PasswordResetDoneView, PasswordChangeView
+
+from account.forms import SignUpForm, SignInForm
+
+logger_email = logging.getLogger(__name__)
+
+
+class SignUp(CreateView):
+    """
+    Отображает и обрабатывает форму регистрации.
+
+    > Авторизованных пользователей перенаправляет в список посещённых городов.
+    """
+
+    form_class = SignUpForm
+    success_url = reverse_lazy('signup_success')
+    template_name = 'account/signup.html'
+
+    def dispatch(self, request, *args, **kwargs):
+        if self.request.user.is_authenticated:
+            return redirect('city-all-list')
+
+        return super().dispatch(request, *args, **kwargs)
+
+    def form_valid(self, form):
+        user = User.objects.create_user(
+            username=self.request.POST['username'],
+            password=self.request.POST['password1'],
+            email=self.request.POST['email'],
+        )
+        user.save()
+        logger_email.info(
+            f"Registration of a new user: {self.request.POST['username']} ({self.request.POST['email']}). "
+            f"Total numbers of users: {User.objects.count()}"
+        )
+        login(self.request, user)
+
+        return redirect('city-all-list')
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+
+        context['page_title'] = 'Регистрация'
+        context['page_description'] = (
+            'Зарегистрируйтесь на сервисе "Мои города" для того, чтобы сохранять свои '
+            'посещённые города и просматривать их на карте'
+        )
+
+        return context
+
+
+class SignIn(LoginView):
+    """
+    Отображает и обрабатывает форму авторизации.
+
+    > Авторизованных пользователей перенаправляет в список посещённых городов.
+    """
+
+    form_class = SignInForm
+    template_name = 'account/signin.html'
+
+    def dispatch(self, request, *args, **kwargs):
+        if self.request.user.is_authenticated:
+            return redirect('city-all-list')
+
+        return super().dispatch(request, *args, **kwargs)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+
+        context['page_title'] = 'Вход'
+        context['page_description'] = (
+            'Войдите в свой аккаунт для того, чтобы посмотреть свои посещённые города '
+            'и сохранить новые'
+        )
+
+        return context
+
+
+def signup_success(request):
+    return render(request, 'account/signup_success.html')
+
+
+class MyPasswordChangeView(PasswordChangeView):
+    template_name = 'account/profile__password_change_form.html'
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+
+        context['page_title'] = 'Изменение пароля'
+        context['page_description'] = (
+            'Для того, чтобы изменить свой пароль, введите старый и новый пароли'
+        )
+
+        return context
+
+
+class MyPasswordResetDoneView(LoginRequiredMixin, PasswordResetDoneView):
+    template_name = 'account/profile__password_change_done.html'
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+
+        context['page_title'] = 'Изменение пароля'
+        context['page_description'] = 'Пароль успешно изменён'
+
+        return context
