+++ account/views/download.py
@@ -1,10 +1,17 @@
+"""
+----------------------------------------------
+
+Copyright © Egor Vavilov (Shecspi)
+Licensed under the Apache License, Version 2.0
+
+----------------------------------------------
+"""
+
 import csv
 import json
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
 from datetime import datetime
 from io import StringIO, BytesIO
-from typing import Literal
 
 import openpyxl
 from django.http import HttpResponse, Http404
@@ -18,12 +25,10 @@ from services.db.visited_city_repo import get_all_visited_cities
 
 class Report(ABC):
     @abstractmethod
-    def __init__(self, user_id: int) -> None:
-        ...
+    def __init__(self, user_id: int) -> None: ...
 
     @abstractmethod
-    def get_report(self) -> list[tuple]:
-        ...
+    def get_report(self) -> list[tuple]: ...
 
 
 class CityReport(Report):
@@ -37,7 +42,13 @@ class CityReport(Report):
         ]
         for city in cities:
             result.append(
-                (city.city.title, str(city.region), str(city.date_of_visit), str(city.has_magnet), str(city.rating)),
+                (
+                    city.city.title,
+                    str(city.region),
+                    str(city.date_of_visit),
+                    str(city.has_magnet),
+                    str(city.rating),
+                ),
             )
         return result
 
@@ -49,27 +60,28 @@ class RegionReport(Report):
     def get_report(self) -> list[tuple]:
         regions = get_all_visited_regions(self.user_id)
         result = [
-            ('Регион', 'Всего городов', 'Посещено городов, шт', 'Посещено городов, %', 'Осталось посетить, шт'),
+            (
+                'Регион',
+                'Всего городов',
+                'Посещено городов, шт',
+                'Посещено городов, %',
+                'Осталось посетить, шт',
+            ),
         ]
         for region in regions:
-            result.append(
-                (region, 10, 5, 50, 5)
-            )
+            result.append((region, 10, 5, 50, 5))
         return result
 
 
 class Serializer(ABC):
     @abstractmethod
-    def convert(self, report):
-        ...
+    def convert(self, report): ...
 
     @abstractmethod
-    def content_type(self):
-        ...
+    def content_type(self): ...
 
     @abstractmethod
-    def filetype(self):
-        ...
+    def filetype(self): ...
 
 
 class TxtSerializer(Serializer):
@@ -139,8 +151,13 @@ class AreaReport(Report):
     def get_report(self) -> list:
         areas = get_visited_areas(self.user_id)
         result = [
-            ('Федеральный округ', 'Всего регионов, шт', 'Посещено регионов, шт',
-             'Посещено регионов, %', 'Осталось посетить, шт'),
+            (
+                'Федеральный округ',
+                'Всего регионов, шт',
+                'Посещено регионов, шт',
+                'Посещено регионов, %',
+                'Осталось посетить, шт',
+            ),
         ]
         for area in areas:
             result.append(
@@ -176,7 +193,9 @@ def download(request):
     else:
         raise Http404
 
-    response = HttpResponse(buffer.convert(report.get_report()).getvalue(), content_type=buffer.content_type())
+    response = HttpResponse(
+        buffer.convert(report.get_report()).getvalue(), content_type=buffer.content_type()
+    )
     filename = f'MoiGoroda__{request.user}__{int(datetime.now().timestamp())}'
     response['Content-Disposition'] = f'attachment; filename={filename}.{buffer.filetype()}'
     return response
