+++ account/views/profile.py
@@ -1,0 +1,48 @@
+"""
+----------------------------------------------
+
+Copyright © Egor Vavilov (Shecspi)
+Licensed under the Apache License, Version 2.0
+
+----------------------------------------------
+"""
+
+from django.contrib.auth.models import User
+from django.shortcuts import get_object_or_404
+from django.urls import reverse_lazy
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.views.generic import UpdateView
+
+from account.forms import UpdateProfileForm
+from services import logger
+
+
+class Profile(LoginRequiredMixin, UpdateView):
+    form_class = UpdateProfileForm
+    success_url = reverse_lazy('profile')
+    template_name = 'account/profile.html'
+
+    def get_object(self, queryset=None):
+        """
+        Убирает необходимость указывать ID пользователя в URL, используя сессионный ID.
+        """
+        return get_object_or_404(User, pk=self.request.user.pk)
+
+    def form_valid(self, form):
+        """
+        Переопределение этого метода нужно только для того, чтобы произвести запись в лог.
+        """
+        logger.info(
+            self.request,
+            f"Updating user's information: {self.request.user.username} ({self.request.user.email})",
+        )
+        return super().form_valid(form)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+
+        context['active_page'] = 'profile'
+        context['page_title'] = 'Профиль'
+        context['page_description'] = 'Просмотр и изменения персональной информации'
+
+        return context
